Android Kotlin Internship at Wirtek
In this internship, we built a podcast listener app in 4 weeks (2 sprints of 2 weeks each). I worked in a team of 2 with my internship colleague and we were  managed by a Senior Android Programmer inside the company that helped us with project management and Jira experience with ticketing and all. We also worked together with a team of 2 IOS interns in order to make the app similar on both type of devices (Android or IOS).